+++ custom_components/growatt_modbus/const.py
@@ -14,6 +14,7 @@ PLATFORMS: list[Platform] = [
     Platform.SELECT,
     Platform.SENSOR,
     Platform.SWITCH,
+    Platform.TIME,
 ]
 
 # Config entry keys